Allow move to CUDA from top-level module
When defining a network in Torch it is customary to move it
to the currently selected GPU through a single call to the
top-level module's cuda() method.

This patch allows this by merely asserting in the type()
method that the target type is 'torch.CudaTensor', rather
that dropping an error unconditionally.

In this patch we are also conveniently moving any added
module to CUDA, since it only makes sense to add CUDA
modules to a DataParallelTable container.